squash: rename
